Moved mapline series into a subfolder.
