+++ js/Series/MapLine/MapLineSeries.js
@@ -21,9 +21,9 @@ var __extends = (this && this.__extends) || (function () {
         d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
     };
 })();
-import BaseSeries from '../Core/Series/Series.js';
-import MapSeries from './Map/MapSeries.js';
-import U from '../Core/Utilities.js';
+import BaseSeries from '../../Core/Series/Series.js';
+import MapSeries from '../Map/MapSeries.js';
+import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
 /* *
  *
