+++ js/masters/modules/map.src.js
@@ -16,7 +16,7 @@ import '../../Mixins/ColorMapSeries.js';
 import '../../Maps/MapNavigation.js';
 import '../../Maps/MapPointer.js';
 import '../../Series/Map/MapSeries.js';
-import '../../Series/MapLineSeries.js';
+import '../../Series/MapLine/MapLineSeries.js';
 import '../../Series/MapPointSeries.js';
 import '../../Series/MapBubbleSeries.js';
 import '../../Series/HeatmapSeries.js';
